services/trade_producer/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["rishikeshF <rishikeshfulari1508@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
quixstreams = "^2.6.0"
websocket-client = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
